+++ src/interactions.js
@@ -1,8 +1,6 @@
 import { arr } from './index.js';
 
-const container = document.getElementById('tasks-list');
-
-export function saveToLocalStorage() {
+function saveToLocalStorage() {
   localStorage.setItem('arr', JSON.stringify(arr));
 }
 
@@ -18,25 +16,4 @@ export function addNew() {
     arr.push(task);
     saveToLocalStorage();
   }
-}
-
-export async function deletingTask() {
-  const { arrayIndex } = await import('./index.js');
-  arr.splice(arrayIndex, 1);
-  arr.forEach((element, index) => {
-    element.index = index + 1;
-  });
-  const { removeAllTasks } = await import('./index.js');
-  const { component } = await import('./index.js');
-  saveToLocalStorage();
-  removeAllTasks(container);
-  component();
-}
-
-export async function editingTask() {
-  const { removeAllTasks } = await import('./index.js');
-  const { component } = await import('./index.js');
-  saveToLocalStorage();
-  removeAllTasks(container);
-  component();
 }
